requirements/base.txt
Django==4.2
Pillow==9.5.0
psycopg2-binary
django-allauth
django-phonenumber-field==7.0.2
python-dotenv
django-ajax-datatable
django-widget-tweaks
django-simple-history
django-dynamic-formsets
drf-spectacular
djangorestframework
markdown
django-filter
numpy